backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
psycopg2-binary
python-dotenv
Flask-Cors